# whatshap haplotag - Speed improvments