--- a/whatshap/cli/haplotag.py
+++ b/whatshap/cli/haplotag.py
@@ -619,7 +619,9 @@ def run_haplotag(
                             alignment.set_tag("PS", value=None)
 
                     bam_writer.write(alignment)
-                    if haplotag_list is not None and not (alignment.is_secondary or alignment.is_supplementary):
+                    if haplotag_list is not None and not (
+                        alignment.is_secondary or alignment.is_supplementary
+                    ):
                         print(
                             alignment.query_name,
                             haplotype_name,