This PR includes some speed improvement to the `haplotag` command. Some improvements only apply to linked reads, e.g. 6954d1fd3f2a714c9e7080783029da84c4869599. 

I benchmarked using a linked-read dataset of about 1M reads.

**Before**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups > /dev/null
  Time (mean ± σ):     12.036 s ±  0.156 s    [User: 11.856 s, System: 0.162 s]
  Range (min … max):   11.810 s … 12.296 s    10 runs
```


**After**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups > /dev/null
  Time (mean ± σ):      9.661 s ±  0.198 s    [User: 9.502 s, System: 0.150 s]
  Range (min … max):    9.239 s …  9.923 s    10 runs
```

That is about ~20% faster !!