@pontushojer 2022-07-20
I included some additional improvement for when using `--ignore-linked-read` in https://github.com/whatshap/whatshap/pull/387/commits/7de7e1e588a8facaf506fdcafd64fd8c8d050e3d and https://github.com/whatshap/whatshap/pull/387/commits/e89657c73b4ca69cc8b2b4730223d1af76deecab. 

Here is the updated benchmarks, now including runs with `--ignore-linked-read`.

## Without `--ignore-linked-read`

**Before**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups > /dev/null
  Time (mean ± σ):     11.916 s ±  0.179 s    [User: 11.751 s, System: 0.157 s]
  Range (min … max):   11.701 s … 12.275 s    10 runs
```

**After**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups > /dev/null
  Time (mean ± σ):      9.739 s ±  0.226 s    [User: 9.578 s, System: 0.152 s]
  Range (min … max):    9.428 s … 10.121 s    10 runs
```
=> 18.3% faster

## With `--ignore-linked-read`

**Before**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups --ignore-linked-read > /dev/null
  Time (mean ± σ):     10.687 s ±  0.180 s    [User: 10.543 s, System: 0.138 s]
  Range (min … max):   10.397 s … 10.963 s    10 runs
```

**After**
```
Benchmark #1: whatshap haplotag chunks/chr22_14029500_18289499.calling.phased.vcf.gz chunks/chr22_14029500_18289499.calling.bam --linked-read-distance-cutoff 30_000 --reference ../chr22_testdata/chr22.fasta --ignore-read-groups --ignore-linked-read > /dev/null
  Time (mean ± σ):      8.462 s ±  0.122 s    [User: 8.325 s, System: 0.130 s]
  Range (min … max):    8.302 s …  8.712 s    10 runs
```
=> 20.8% faster