@marcelm 2022-09-06
Thanks again! A nice improvement :-)

(I forgot that we need to add an entry to the changelog. I’ll merge first and then do that in a separate commit.)